refactor(vite-plugin-remix): collapse /runtime + /ui into /client
Single subpath export so the consumer reaches both helpers from the
same import root:

  import { boot, Script } from 'vite-plugin-remix/client'

File layout mirrors the surface — boot.ts (was runtime/browser-entry.ts)
holds the runtime kick-off, script.tsx keeps the SSR <Script>, and
client/index.tsx re-exports both. Drop the /runtime and /ui sub-exports
from package.json.

Update my-remix-app and the two READMEs to import from
vite-plugin-remix/client.
