+++ js/package/vite-plugin-remix/src/client/boot.ts
