+++ js/package/vite-plugin-remix/src/client/index.tsx
@@ -1,0 +1,5 @@
+export { boot } from './boot.ts'
+export type { BootOptions } from './boot.ts'
+
+export { Script } from './script.tsx'
+export type { ScriptProps } from './script.tsx'
