+++ js/package/vite-plugin-remix/src/client/script.tsx
